+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log (vs-media-player)
 
+## 1.1.0 (July 30th, 2017; select output devices)
+
+* can select output devices via `Media Player: Select output` (`extension.mediaPlayer.selectPlayerOutput`) command now
+
 ## 1.0.6 (July 30th, 2017)
 
 * first stable release